Put all program files into folders, added abouts
Folders:
Calculator Apps - Geometric Calculator & calculator
Codecademy
Guessing Game
Rock paper Scissors
