+++ Codecademy/about.txt
@@ -1,0 +1,1 @@
+These are all files used for extension of codecademy course files and backups.
